requirements.txt
Django==3.1
python-telegram-bot==12.8
XlsxWriter==1.3.3
psycopg2==2.8.5
django-crontab==0.7.1
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.12.0
phonenumbers==8.12.15
Pillow==8.1.0
django-cors-headers==3.6.0
python-dateutil==2.8.1

django-extensions==3.1.0

# Celery
celery~=4.4.0
django-celery-results==1.1.2
django-celery-beat==2.2.0
redis==3.5.3
